requirements.txt
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
mysql-connector-python==8.4.0
openai==1.30.1
httpx==0.27.0
python-dotenv==1.0.1